requirements.txt
pandas==1.5.3
requests==2.28.1
psycopg2==2.9.6
python-dotenv==0.21.0
numpy==1.24.1

